LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-is-still-set-when-reporting-exception-onerror-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: foo


PASS window.onerror handler restores window.event after it reports an exception
